+++ src/App.js
@@ -6,8 +6,8 @@ import SearchUser from './pages/SearchUser/SearchUser.jsx';
 import Join from './pages/Join/JoinEmail/JoinEmail';
 import JoinProfileSetting from './pages/Join/JoinProfileSetting/JoinProfileSetting';
 import MyProfile from './pages/Profile/MyProfile/MyProfile';
-import Followers from './pages/Profile/Followers';
-import Followings from './pages/Profile/Followings';
+import Followers from './pages/Follow/Followers/Followers';
+import Followings from './pages/Follow/Followings/Followings';
 import ProfileModification from './pages/ProfileModification/ProfileModification';
 import AddProduct from './pages/Product/AddProduct';
 import ProductModi from './pages/Product/ProductModi';
